Move Reverse to WaveTrackUtilities
